app/requirements.txt
fastapi
uvicorn
sqlalchemy
passlib
bcrypt
python-jose
python-multipart
python-dotenv
# driver interact with db in async manner
asyncmy
